Arreglo de los angulos en el editor y UI cliente.
Archivo yaml generado con mas presicion (en cuanto a los angulos).
Verificaciones UI cliente sobre pantalla de seleccion de sala.

+++ editor/src/conversor.cpp
@@ -18,7 +18,7 @@ QPair<float,float> Conversor::a_metros(const QGraphicsPixmapItem &obj, float ang
     return QPair<float,float>(x_centro, y_centro);
 }
 
-float Conversor::a_radianes(float angulo) const
+double Conversor::a_radianes(float angulo) const
 {
     return qDegreesToRadians(angulo);
 }
@@ -27,14 +27,3 @@ float Conversor::pixel_a_metros(float pixel) const
 {
     return pixel / ESCALA;
 }
-
-// QPointF Conversor::a_pixeles(const QPair<float,float> pos) const
-// {
-//     float x, y;
-//     float radio = ancho/2;
-//     x = pos.first - alto/2 * qSin(angulo) - radio * qCos(angulo);
-//     y = pos.second - alto/2 * qCos(angulo) + radio * qSin(angulo);
-//     x *= ESCALA;
-//     y *=  ESCALA;
-//     return QPointF(x, y);
-// }
